Split NOTICE in to NOTICE, LICENSE
The text of the Apache License is not Copyright AOSP as notice above
the license would imply.

Thus, this separates the license from their notice, and returns the
instructions to the license as the license exists in the original form.

Also removes the empty MODULE_LICENSE_APACHE2 marker file.
The MODULE_LICENSE_APACHE2 file seems to be specific to Android Open
Source Project tooling, as can be read about here:

https://web.archive.org/web/20150523042805/http://www.dejacode.org/android_case_study.html

Signed-off-by: Eric Herman <eric@freesa.org>