No longer leak the sparse file backing buffers.
Freeing backing buffers of sparse files are the responsibility of the
caller. As such, a reference must be held in order to free them.

Adds a vps_list of "long_life_bufs", and passes it to functions which
create backing buffers for sparse files.

Also adds the super-block block_data to the "long_life_bufs".

At the end of verbose test run we can see:
"Freed 4239504 bytes of long_life_bufs"

And valgrind reports:
    in use at exit: 0 bytes in 0 blocks
  total heap usage: 66 allocs, 66 frees, 6,400,459 bytes allocated

  All heap blocks were freed -- no leaks are possible
  ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)

With all leaks gone from the current tests, updated the Roadmap.

Signed-off-by: Eric Herman <eric@freesa.org>